+++ styles/helvetica/boxes.scss
@@ -222,3 +222,11 @@
   vertical-align: rem(1px);
   margin-left: 0.5em;
 }
+
+.subscriptions-request-alert {
+  padding: 0.3125rem 0.625rem;
+  border: 1px solid #bce8f1;
+  border-radius: 4px;
+  background-color: #d9edf7;
+  color: #31708f;
+}
